Time: 3 ms (71.91%), Space: 5.8 MB (95.25%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        long mid;
+        long long mid;
         int l = 0, r = x, ans;
         while(r >= l){
             mid = r + l >> 1;
